requirements.txt
aesara==2.9.4
aiohttp==3.10.5
albumentations==2.0.5
amdsmi==6.3.3
AppKit==0.2.8
asgiref==3.8.1
astunparse==1.6.3
attr==0.3.2
av==14.2.0
beartype==0.20.0
beautifulsoup4==4.13.3
bokeh==3.6.0
boto3==1.37.13
botocore==1.34.69
Brotli==1.0.9
brotlicffi==1.1.0.0
cairocffi==1.7.1
Carbon==1.1.10
cartopy==0.24.1
cftime==1.6.4.post1
clip==0.2.0
comet_ml==3.49.5
conda==4.3.16
ConfigParser==7.2.0
contextlib2==21.6.0
coremltools==8.2
cryptography==44.0.2
ctypes_snappy==1.03
Cython==3.0.12
deeplearning==0.0.1
defcon==0.12.0
defusedxml==0.7.1
diffusers==0.32.2
dill==0.3.8
dl==0.1.0
docutils==0.18.1
docutils==0.18.1
dvclive==3.48.2
etcd==2.0.8
expecttest==0.3.0
fastcluster==1.2.6
fastparquet==2024.11.0
flatbuffers==25.2.10
flint==0.0.1
Foundation==0.1.0a0.dev1
freetype_py==2.5.1
gdown==5.2.0
genapi==0.0.8
gitdb_speedups==0.1.0
glyphsLib==6.10.1
gmpy2==2.2.1
h2==4.2.0
h5py==3.11.0
h5py==3.12.1
HTMLParser==0.0.2
hub_sdk==0.0.18
hypothesis==6.129.3
ini2toml==0.15
ipykernel==6.29.5
ipython==8.27.0
ipython==8.12.3
ipywidgets==7.8.1
isal==1.7.2
jnius==1.1.0
kerchunk==0.2.8
keyring==24.3.1
lap==0.5.12
lmdb==1.4.1
lockfile==0.12.2
lvis==0.5.3
lxml==5.2.1
lz4==4.3.2
lzmaffi==0.3.0
matchpy==0.5.5
mayavi==4.8.2
mct_quantizers==1.5.2
MNN==3.1.1
mock==5.2.0
model_compression_toolkit==2.3.0
monkeytype==23.3.0
moviepy==2.1.2
mss==10.0.0
mtrand==0.1
multiset==3.2.0
munkres==1.1.4
mypy==1.11.2
ncnn==1.0.20241226
nncf==2.15.0
nose==1.3.7
numarray==1.5.1
Numeric==24.2
numexpr==2.8.7
odfpy==1.4.1
olefile==0.47
onnx2tf==1.27.0
onnxscript==0.2.2
onnxslim==0.1.48
openpyxl==3.1.5
openvino==2025.0.0
optree==0.14.1
paddle==1.0.2
pafy==0.5.5
paramiko==2.8.1
pickle5==0.0.12
pillow_heif==0.22.0
pkg1==0.0.3
pooch==1.8.2
psycopg2==2.9.10
pycocotools==2.0.8
pycosat==0.6.6
pydot==3.0.4
pydy==0.7.1
pygit2==1.17.0
pyglet==2.1.3
pygraphviz==1.14
PyInstaller==6.12.0
pymc==5.21.1
pymc3==3.11.6
pynvml==12.0.0
pyodide==0.0.2
pyOpenSSL==24.2.1
pyOpenSSL==25.0.0
PyQt4==4.11.4
PyQt5==5.15.11
PyQt5_sip==12.13.0
PyQt6==6.8.1
pysat==3.2.1
PySide2==5.15.2.1
PySide6==6.8.2.1
pytest==7.4.4
pytest.egg==info
pytest_timeout==2.3.1
python-dotenv==1.0.1
pytorch_lightning==2.5.0.post0
pytubefix==8.12.2
pywin32==305.1
pyxlsb==1.0.10
QtPy==2.4.1
Quartz==0.0.1.dev0
railroad==0.5.0
ray==2.43.0
redis==5.2.1
reportlab==4.3.1
requests_kerberos==0.15.0
Res==0.1.7
roboflow==1.1.58
s3fs==2024.6.1
safetensors==0.5.3
sage==0.0.0
scikit_learn==1.6.1
scikits.talkbox==0.2.5
selenium==4.29.0
sentry_sdk==2.18.0
setuptools_scm==8.2.0
Shapely==2.0.7
shiboken2==5.15.2.1
shiboken6==6.8.2.1
simplejson==3.20.1
sip==6.7.12
slack_sdk==3.34.0
smbprotocol==1.15.0
sony_custom_layers==0.3.0
Sphinx==8.2.3
SQLAlchemy==2.0.34
statsmodels==0.14.2
streamlit==1.37.1
super_gradients==3.7.1
symengine==0.14.0
tabulate==0.9.0
tensorflowjs==4.22.0
tflite_runtime==2.14.0
tflite_support==0.4.4
theano==1.0.5
thread==2.0.5
threadpoolctl==3.6.0
tomli_w==1.2.0
torch_xla==2.6.0
torcharrow==0.1.0
torchaudio==2.6.0
torchrec==1.1.0
tornado==6.4.1
traitlets==5.14.3
tritonclient==2.55.0
trove_classifiers==2025.3.13.13
truststore==0.10.1
ufoLib2==0.17.1
uharfbuzz==0.46.0
ujson==5.10.0
unicodedata2==15.1.0
urllib3_secure_extra==0.1.0
wandb==0.19.8
wmi==1.5.1
x2paddle==1.6.0
xarray==2023.6.0
xattr==1.1.4
xlrd==2.0.1
xlsxwriter==3.2.2
xmlrpclib==1.0.1
xmlrunner==1.7.7
xx==3.3.2
yarl==1.11.0
yt_dlp==2025.2.19
z3==0.2.0
zopfli==0.2.3.post1
zstandard==0.23.0
torch~=2.4.1
flask~=3.0.3
pillow~=10.4.0
Flask-Cors~=5.0.0
opencv-python~=4.11.0.86
numpy~=1.24.4